add scripting for player input to trigger failure
A new function triggers an animation in the `FluffyRagdoll` and freezes
user input.  This function is called upon user input with a parameter
used to determine which "fall over" animation it should use.

As a placeholder, if the user presses a key for which an animation is
not yet available, Fluffy simply flips horizontally.

Whether or not the `decide` function signals "won" or "lost" now depends
on whether the game has detected a key pressed (i.e. the `key_pressed`
variable is set to `true`)