add a "jump" button to the Fluffy Ragdoll game
Since pressing up or right pushes Fluffy over to the right and pressing
down or left pushes Fluffy over to the left, I thought it would be
funny if pressing space caused a really cheesy jump animation to play.
Experimentation has shown that it is indeed funny at least to me, so in
the game it shall stay.