move some elements around in the in-game HUD
The label has been renamed from GameInstruction back to Label to reflect
its actual purpose as a general-purpose text box and has been centered
in the HUD mostly for logical reasons.

The visibility of Label and the TimerProgress progress bar have been set
to false because that is how they will be when the game is initially
started up.  They will be re-activated when a minigame starts.