fix message scroll speed during "zoom" animation
I added a constant `MESSAGE_ZOOM_DX` so I can easily edit how quickly
the messages scroll.  I got my reasoning backwards when I initially
scripted the animation; it needs to go twice as fast as the screen width
per second, so it only takes half a second for the message to appear and
disappear.