add a scene to draw minigames on.
The `MinigameCanvas` class is used to handle minigames now instead of
`Main`.  The reason for this is twofold: One, it's the only way I could
think of to get the HUD to draw on top of the minigame, rather than
beneath it; and two, it helps to push a lot of the logic of swapping and
discarding minigames into a separate class, leaving `Main` to just
handle meta-game stuff like lives and streaks.