move saw minigame win condition check.
The win condition on the saw minigame was placed outside of the check
that the minigame's controls have been locked and the minigame was
constantly emitting a win condition, causing the game to lock up.  This
commit fixes that by only checking for the win condition when the
controls are unlocked.