add a background image to the saw minigame.
A background image (actually a foreground image made to look like a
background image) has been drawn and added to the saw minigame.  Several
in-game elements have been moved around to better fit within the image,
and a white background has been placed behind the "main play area"
behind the saw & logs to color in the one transparent spot that was made
for the minigame.