setup one-time animation for Explosion.
The Explosion class now has a rudimentary script which causes it to
remove itself from the queue as soon as its own animation finishes.  The
animation itself has been set to 20 frames per second and will not loop,
allowing the `animation_finished` signal to be thrown.  Additionally,
the Explosion has been quadrupled in size to get the nice pixelated look
I wanted.