add hit detection to the Anchovy.
A rudimentary collision detection script has been added for the Anchovy.
This will detect if the Anchovy comes into contact with a Pizza and
emit the "hit" signal if so.  It also shows the "Boom" sprite.

The Anchovy has also been changed from an Area2D to a RigidBody2D, since
this will make it easier to do the Asteroids-style, momentum-based
movement I'm looking for.