fix collision detection layers & masks.
YOU GUYS I TOTALLY DIDN'T KNOW HOW LAYERS AND MASKS WORK IN GODOT THIS
SIMPLIFIES MY LIFE SO MUCH!!!!!!!!

Here's how the collision layers and masks are going to work for this
minigame:

The Player is ON collision layer 1 and SCANS layer 2.

The Pizzas are ON collision layer 2 and SCAN no layers.

Pepperonis are ON collision layer 3 and SCAN layer 2.

The basic idea is that the player and the pepperonis all scan for pizzas
and handle how they interact with the pizzas themselves.