correct collision detection data for Lose.
After a surprisingly reasonable amount of trial and error, I've figured
out the optimal solution to make a rudimentary platformer in the engine.
The FluffyPlatformer scene has been changed to a RigidBody2D and the
BasicFloor has been changed to a KinematicBody2D.  The BasicFloor's
CollisionShape2D has also been rotated 180 degrees to make the one-way
collision work.

Additionally, the Gravity Scale on the FluffyPlatformer has been
adjusted to 3, so that he'll fall at a more satisfying speed.