[ZA] add a basic script for the minigame.
The initial script has a basic test mainloop which slices the 'za on a
press of the spacebar.  A "won" signal is declared when the 'za is fully
eaten, but if the minigame is asked to `decide` early it will emit the
"won" signal if the 'za is fully sliced; otherwise a "lost" signal is
emitted.