[ZA] add code to check rotation & attempt a slice.
The `try_slice` function in ZaMinigame will check the rotation of the
slicer against the rotation of the slice guide.  If they are within a
certain margin of error of each other, the slice is done and the number
of slices is incremented.  Otherwise, the player is declared a loser.

The function wraps the rotations of both the slicer guide and the slicer
itself to try and ensure that the guide can be attacked from either
side.

This function has been implemented in the minigame's mainloop.