[ZA] reduce slicer's speed to 90 degrees/sec.
I was worried that the slicer wouldn't move fast enough to perform all
four slices in the four seconds provided, but it turns out it's more
than doable.

Since the margin of error is so tight, the speed of the slicer's
movement was actually a liability, so halving it like this is super
helpful.