[MASTER] simplify the minigame queue to an array.
Rather than having to export the minigames one by one and then create
IDs for each of them, the minigame queue is now an array in
`MinigameCanvas` named `Minigames` which contains a number of `preload`
statements which produce `PackedScene` objects which we can `instance`
to get the minigames.

This change means that we will now only need to modify the `Minigames`
queue in one file rather than editing three.

The `reset_arrays` function in `Main` has been adjusted to reflect this
change.