[META] remove the "vote" minigame.
One of my core design principles is don't get political in an
entertainment piece.  The vote minigame wasn't just too political, it
wasn't even that good of a joke and it wasn't fun to play, so that's
three strikes against it, I'm gonna remove it.  May it never darken our
doors again.