[SHOOTER] add a function to respond to kills.
When an enemy is killed, they will emit the "exploded" signal.  At some
point along the line that signal has to get to the `ShooterMinigame`,
and at that point the `_on_enemy_exploded` function will instance an
explosion at the specified location.

This function will also increment the kill counter, and if it gets to
the constant `KILLS_REQUIRED`, currently set to 3, it will `stop` the
minigame and emit the "won" signal.