[SHOOTER] quadruple enemy speed.
Because I couldn't get the enemies to kill the player, the challenge
factor for this minigame has shifted from kill-or-be-killed to actually
killing harmless ships (we're the good guys, right?) so the enemies have
had their speed greatly increased to give them a fighting chance.

Testing has not revealed any problems with collision detection at this
time.