[SHOOTER] initialize the Star node.
The Star is a 2x2 pixel `ColorRect` which chooses a random brightness
when it spawns.  It zips down to the bottom of the screen and then
deletes itself when it reaches the bottom.

The star's speed is a function of its brightness, so dimmer stars which
look farther away will move slower.