[SHOOTER] initialize BigBoom script.
The initial draft of the BigBoom script has been added.  When ready, it
will first adjust the position of the ColorRect to try and line it up
with the position of the BigBoom itself, then as soon as its process
begins it will start to fade away.

When the transparency goes below the `COLLISION_ALPHA` constant,
initialized to 0.8, it will stop colliding with enemies, thus causing
them to stop dying.  Then when its alpha reaches 0, it will dequeue
itself.