[SHOOTER] completely change how the missile works.
After a hair-pulling amount of testing, I have determined that making
the missile's explosion a separate entity from the missile itself was
causing way too many problems and I have decided to change how it works
by having the missile's collision shape expand in order to destroy the
enemy ships.  The BigBoom node has been removed entirely.

A lot of work went into this commit and it would be arduous and not
worth anyone's time to attempt to do atomic changes like you're supposed
to do with git commits, sorry.