[SHOOTER] add new BigBoom animation.
The BigBoom animation is now done by revealing a ColorRect that's
attached to the missile itself and fading it out in the missile's
process.  The missile dequeues after the BigBoom fades completely, and
before that the collision box is disabled by a timer.

The missile's collision shape is now what destroys enemies, by inflating
to a radius of 120, which fills the screen.  This thing is so OP it's
not even funny.