[MASTER] add a box to go around the instructions.
The `InstructionLabel` now has a box which is drawn around it using
`ColorRect` objects.  Originally I was going to do it with theming but
it turns out that in order to do that I would have to have an entire
Godot GUI theme embedded in the object and frankly fuck that.

The positions associated with the "center" of the screen have been
adjusted to recenter the instruction label in-game.  I still don't know
what's causing that bug where it appears in the middle at the start of
the game.