[MASTER] initialize the Master scene.
The Master scene will be used for the game's menus and to control the
Main scene.  It is now the new main scene for the project.

A function has been added to Master which sets the version number at the
start of the game.

Another function has been added which starts a new game when Main is
ready.