[MASTER] add a script for the LegalStuffDialog.
The Legal Stuff dialog now has scripting for the buttons that it
contains.

The "Read the Unlicense" button will toggle the full text of the
unlicense and a link to the Unlicense website.

The URLs to the Godot Engine license and the Unlicense will emit a
signal instructing a separate dialog box to warn the user they are
attempting to open an external URL.