[MASTER] add game modes.
The `new_game` function in `Main` now takes a parameter `mode` which
sets the game mode, defined in the enum `GameMode`.  Various aspects of
`Main` have been tweaked to behave differently according to the current
game mode.

If `mode` is `GameMode.DEBUG`, `new_game` will take an additional
parameter `id` which tells it which minigame is being debugged.