[MASTER] rearrange main menu buttons.
The buttons on the main menu have been reformatted to gravitate toward
the bottom of the screen.  They have also been resized so that they will
be substantially larger.  The `NewGameButton` node has had its `Label`
moved down so that it will not overlap with the text of the button.