[META] add a stylebox for the instruction box.
The `InstructionLabel` in `InGameHUD` now gets its own very special
stylebox that has all the stylings I really wanted to see out of it.

`InGameHUD` itself now uses the default Minigame Madness theme and its
visibility is now set in `Main` rather than in `InGameHUD` because
frankly I was getting sick and tired of having to toggle it every time I
wanted to make an edit.