[META] move master-specific themes/fonts to master.
The themes, styleboxes, and fonts exclusive to objects in Master have
been moved to Master.

Normally that would be all of them, but the "Minigame Madness" theme is
meant to be the default theme for the project, so it makes more sense to
keep it in the Shared folder, and the "generic" fonts (halogen,
heavy_equipment, janitor, and modern_DOS) have been kept in there as
well because likewise they are meant to be shared resources.