[SHARED] re-implement projectiles as kinematic.
As a reflection of previous changes to the game to improve collision
detection, projectiles will now be `KinematicBody2D` objects.

Code has been added to the projectile script to handle movement and
signal collisions, and an exported variable `_disappear_on_collision`
has been added which defaults to true which will determine if the
projectile disappears on detecting a collision.