[SHOOTER] reset the missile's hitbox on game start.
If the shooter minigame is played more than once in a single session of
Minigame Madness (for example, the player starts a new game) the
missile's hitbox isn't reset to its original size and remains as large
as it was when it exploded in the previous game, leading to premature
detonation.  This commit resolves that issue by having the missile set
its collision shape's radius to 5.909 in the `_ready` function.