[SHOOTER] lower boundary of minigame to bottom.
The red boundary around the shooter minigame is noticeably higher than
the actual lower boundary of the game screen, which is at roughly y
coordinates 460.0 owing to the timer at the bottom.  Since enemies and
stars passing this boundary is a noticeable oversight, lowering the
boundary makes this implicitly a part of the game window even if the
player can't do anything about that area of the screen.