[SHOOTER] expand game boundary to 460 px.
The shooter minigame is too easy.  There just isn't enough room for the
enemy ships to have a realistic chance of maneuvering out of the way.
To help make up for this, the game's boundary has been extended to 460
pixels in length and more paths have been added for the enemy ships.
It's still not a very difficult game, but it's no longer as trivial as
it was before.