[SHOOTER] remove unnecessary key prompts.
With the new screen size, the key prompts were getting in the way, and
since the controls for the shooter minigame aren't exactly unintuitive I
figured it would be best to just delete them rather than trying to
rearrange things so they'd be out of the way.

The missile key, however, has been kept, and has been moved off to the
side of the screen, since this is the one and only clue to the player
that they have a missile weapon at all.