[META] remove unnecessary global classes.
A lot of the early code declared scripts as global classes because I
wasn't familiar with GDScript.  This resulted in a lot of scripts being
declared globally that didn't need to be.

This commit removes the global status of all of the classes that don't
need to be global.  The only classes which remain global are Minigame
and Projectile2D, which are both inherited multiple times throughout the
project.