[MASTER] add nightmarish godot licensing info.
The licensing information has been modified to be more obvious and more
comprehensive.  A new window dialog has been created specifically
dedicated to the frankly ridiculous number of third-party licenses which
the Godot Engine requires.

This totally ludicrous change was made in order to be fully compliant
with all licenses that the Godot Engine is subject to, so that Minigame
Madness binaries can be safely distributed without fear of causing
issues for myself or my end-users.