[PANTS] set up the pants & win condition.
The Pants scene has been written, with relevant scripting to make them
fall and put them on the player when they collide.

Collision detection information has been configured for both the pants
and the player.

The "Put on Pants!" minigame's script has been configured to spawn pants
along a certain path and to detect the game's win condition.

Two new sprites have been added to be randomly chosen for the pants
which spawn.