[MASTER] update version number to Alpha 0.11
I'm adopting a more sensible version numbering system for Minigame
Madness, using a Release.Minor(.Patch) format.  See the comment line in
`master/scripts/master.gd` for details.  The first version number to use
this system will be 0.11-alpha

This commit also updates the number of minigames in the readme.