[MASTER] reduce number of lives to 3.
This was something I had intended to do for version 21.18.3 but forgot
to do before announcing the release on Screenshot Saturday.

Because only about half of the planned minigames have been implemented,
I wanted to reduce the number of lives from 5 to 3 temporarily in order
to create a more challenging experience for the player (the player could
safely fail about half of the minigames, which doesn't make for a very
challenging experience).

The life bar has been adjusted and new textures added to show a maximum
of 4 lives (including one bonus life to be implemented).