[META] move the clouds from WIND to SHARED.
The clouds from the "Don't Fall Down!" minigame have been moved to the
shared folder for incorporation into another minigame.

Because the clouds are not a scene in their own right, this consists of
moving the cloud images from `mini/wind/images/backgroud` to
`shared/images/background` and the script from
`mini/wind/scripts/scenery` to `shared/scripts/background`.