[SHOOTER] increase required kills to 5.
Three kills for the "Shoot Three!" minigame is simply not enough to
provide a legitimate challenge; in fact it's so few that the player
could conceivably get all of the required kills with one shot of the
dragon missile or by spamming shots across the screen.  Increasing the
kill count to 5 makes it less likely, though still not impossible, that
the player will manage to do this.