[PITFALL] add win & loss conditions.
A win zone has been added which triggers a win state if it detects the
player entering it.  Additionally, the player will now detect the kill
zone and emit a "died" signal which triggers the loss state for the
minigame.

The collision layers and masks for several in-game objects have been
adjusted in order to facilitate this change.