[PITFALL] add the gators to the minigame.
The "Get Across!" minigame will now randomly choose between a regular
lake with lily pads and a lake with gators in it when the game starts.
Based on its choice, it will perform setup when the minigame is ready
and start the gators' timers when the game starts, if necessary.