[PITFALL] fix player death using states.
Rather than the `_victory_march` variable, the player will now use a
state `States.VICTORIOUS` to determine whether to do the victory march.
The player's default state is `States.DEFAULT`.

If the player dies, their state will be set to `States.DEAD`, which will
disable further state changes and all movement functions.  This will
prevent the player's sprite from being changed after the game ends or
the player somehow dying more than once.

Also in this commit, the player's position is adjusted upon death to
make it appear that their hat falls to the ground.  Gravity is divided
by 2 during this state to emphasize the effect.

This commit intentionally does not disable jumping during the victory
march, because why would you deny the player the joy of doing a little
victory jump?