[PITFALL] implement the pit & vine hazard.
A new hazard `Hazards.PIT` has been added which replaces the pond with a
pit and adds a swinging vine above it.  The player must swing across the
pit on the vine in order to succeed.

I'm not gonna lie, this is a total hackjob and I honestly can't believe
that I got this to work.  The process by which I managed to do this is
so tedious that it's probably not even worth documenting in detail.