[BIGRIGS] add movement for the "2nd player" rig.
The second truck in the "Race that Rig!" minigame will now move when the
game starts and continue moving until the end of the minigame.  Although
the rig not moving was originally part of the joke, people seemed to
find this confusing.

The rig will move at a constant speed until the end of the minigame
(which should coincide with the rig reaching the end of the track),
mimicking the behavior of the AI truck from Big Rigs: Over the Road
Racing.