[PANTS] adjust distribution of the falling pants.
Rather than offsetting the pants spawner by a random number from 0.0 to
1.0 (the far left and far right) it will be offset by a random number
from 0.25 to 1.5.

This change is being made because the average output from a random range
of 0.0 to 1.0 will be 0.5, which is right in the middle of the play
field.  This means that in a disproportionate number of games, the
player can win by not moving and waiting for pants to come to them.

This change also means that pants are more spread out, because the
minimum amount of change is a quarter of the screen away rather than
directly next to the previous spawn point of the pants.