[META] simplify implementation of space bar labels.
The space bar key prompt now implements a built-in label and a script
which extends the key prompt script to simplify the process of labelling
the spacebar in the editor.  Rather than having to add a label manually,
the editor can now set the `_label` variable which updates the text of
the label automatically.

This change was done in order to simplify making the fonts of all of the
in-game keyboard labels consistent and change said font to Halogen in
order to fit with the game's branding.