[DANCE] implement score announcer.
A label which pops up a message to give player feedback on their
accuracy has been added.  Depending on how near the target the player
is, the label will pop up a message which will indicate what score the
player will get from an arrow, from this list:

0 - MISS (gray)
1 - OK (blue)
2 - GOOD! (orange)
3 - GREAT!! (green)
4 - PERFECT!!! (red)

The label is color-coordinated so that different color messages come up
depending on the player's score.