[SHOOTER] have bad guy handle player collision.
It transpires that the bad guy is, by necessity, an Area2D collision
object, rather than a physics body, so the source of the collision
detection has to come from it rather than the player.  This commit
changes functions and collision information around to make this work.

The player's `_on_Player_body_entered()` function has been renamed to
`blow_up()` to make this work.  Obviously, the function will no longer
prompt the bad guy to explode, as it is now happening the other way
around.