[SHOOTER] don't interact w/ player after death.
Previously, the process of stopping the "Shoot Five!" minigame would
always cause it to disable the player's controls, which caused a crash
if the minigame attempted to do this after the player had been removed
due to death.

This commit adds a check to see if the player has died, and only stops
the player if it hasn't.