[PITFALL] disable the vine after grab.
The vine's signals are now disconnected after the player grabs it,
preventing the player from grabbing it twice in a single run.

This change is being made to prevent the vine from "catching" the player
after the player tries to jump off the vine, which can kill some runs by
forcing the player to continue swinging, sometimes until the timer runs
out.