[META] add an input map for quitting the game.
Using the "ui_cancel" input map to quit the game conflicts with
recent changes allowing the use of controller button 1 for the
"action" button.  This commit fixes that by introducing a "quit" input
map, assigned to the Select button and the Escape key, which is used to
quit the game instead.