[SHOOTER] only dequeue the missile once.
A bug previously existed which caused the dragon missile in the "Shoot
Five!" minigame to attempt to dequeue after the player dies, even if it
had already been fired and had been destroyed.  The missile now only
attempts to dequeue if it has not yet been fired.  If it has been fired,
it will be allowed to continue on its course until detonation.