[MASTER] lock controls after quitting.
The `game_in_progress` boolean was previously not set after the game
ended, only after a minigame ended.  Because of this, it was possible to
keep the controls unlocked after the game has concluded, causing it to
be possible to try to quit the game before it begins, which created
problems when the game tried to clear data that didn't exist.

This commit fixes that by setting `game_in_progress` to false in the
`game_ended` function, regardless of why that function was called.