[META] normalize arrow sprites.
Since all of the arrows in the game are identical, they should all use
the same sprite.  The white arrow sprite from the "Dance, Dance, Dance!"
minigame has been moved into the "shared" folder to act as the one and
only true arrow sprite.

Like in the "Dance, Dance, Dance!" minigame, this arrow sprite will be
modulated when different colors are needed.

The arrow key sprites now use this sprite for their arrows.

The arrow buttons in the Controls dialog have not been changed because
they don't preserve their rotation and thus aren't able to use the same
fix.