[BALL] disable ball collision.
Balls will no longer collide with balls.

This is due to an unfortunate problem with Godot's standard collision
detection, which causes the balls to sometimes freeze in place, causing
havoc with the game's playability.

This is also a balance fix, because allowing the balls to bounce off of
one another did undo my attempts to make the minigame more fair.  Oh,
well.