[SHARED] replace sky texture.
The "sky.png" image file has been removed and replaced with a
`GradientTexture` that achieves the same effect much more smoothly.
The colors on this texture are also different, chosen specifically
because the gradient between them is much smoother without a nasty
brown-ish hue appearing in the middle due to errors in the way colors
are mixed on computer systems.