[SHARED] Allow SkyBackground to use a texture.
The `SkyBackground` class now allows the user to import a texture, which
will be used for the sky.

Rather than being a debugging fallback, the black rectangle that was
used before is now the default in the event that a texture is not
submitted.  This will be useful for rendering night skies.