[MASTER] Add new Legal Stuff dialog.
The new Legal Stuff dialog pops up when the "Legal Stuff" button is
pressed in the menu.  The old Legal Stuff dialog, which is now the
Open-Source Licenses dialog, pops up when the "Open-Source Licenses"
button is pressed in the new Legal Stuff dialog.

I will not be taking questions.